moving logput button into original navbar, remove comment stuff, make home component
