+++ client/src/components/Register.js
@@ -188,11 +188,11 @@ class Register extends Component {
                         </p>
 
                         <div className="form-group input-group">
-                        {/* <Link to='/somewhere'> */}
+  
                             <button  type="submit" className="btn btn-primary">
                                 SIgn Up
                             </button> 
-                            {/* </Link> */}
+                       
                         </div>
                         
                     </form>
@@ -217,8 +217,6 @@ class Register extends Component {
 }
 
 
-
-
 Register.propTypes = {
     registerUser: PropTypes.func.isRequired,
     auth: PropTypes.object.isRequired
@@ -232,105 +230,3 @@ const mapStateToProps = state => ({
 export default connect(mapStateToProps,{ registerUser })(withRouter(Register))
 
 
-
-
-
-
-// <!-- Modal -->
-// <div className="modal fade modal-login" id="login-modal" tabindex="-1" role="dialog"
-// aria-labelledby="exampleModalLongTitle" aria-hidden="true">
-// <div className="modal-dialog modal-md" role="document">
-//     <div className="modal-content">
-//         <div className="modal-header">
-//             <h5 className="modal-title" id="signin-brand">Org Spaces</h5>
-//             <button type="button" className="close" data-dismiss="modal" aria-label="Close">
-//                 <span aria-hidden="true">&times;</span>
-//             </button>
-//         </div>
-//         <div className="modal-body">
-//             <form name="login" className="form-group" method="post" action="/login">
-//                 <label for="email">Email Address</label>
-//                 <input type="text" name="email" className="form-control"  placeholder=" Email Address">
-//                 <label for="password">Password</label>
-//                 <input type="password" name="password" className="form-control" autocomplete="on"  placeholder="Password">
-            
-//             <!-- showing error to client -->
-//             <div className="message-login"></div>
-//         </div>
-//         <div className="modal-footer">
-//         <button type="reset" className="btn btn-secondary" id="close" data-dismiss="modal">Close</button>
-//         <input type="submit" className="btn btn-primary" id="login-button" value="Login" /> 
-//         </form>
-//         </div>
-//     </div>
-// </div>
-// </div>
-
-
-
-
-
-{/* <div className="container" style={{ marginTop: '50px', width: '700px'}}>
-            <h2 style={{marginBottom: '40px'}}>Registration</h2>
-            <form onSubmit={ this.handleSubmit }>
-                <div className="form-group">
-                    <input
-                    type="text"
-                    placeholder="Name"
-                    className={classnames('form-control form-control-lg', {
-                        'is-invalid': errors.name
-                    })}
-                    name="name"
-                    onChange={ this.handleInputChange }
-                    value={ this.state.name }
-                    />
-                    {errors.name && (<div className="invalid-feedback">{errors.name}</div>)}
-                </div>
-                <div className="form-group">
-                    <input
-                    type="email"
-                    placeholder="Email"
-                    className={classnames('form-control form-control-lg', {
-                        'is-invalid': errors.email
-                    })}
-                    name="email"
-                    onChange={ this.handleInputChange }
-                    value={ this.state.email }
-                    />
-                    {errors.email && (<div className="invalid-feedback">{errors.email}</div>)}
-                </div>
-                <div className="form-group">
-                    <input
-                    type="password"
-                    placeholder="Password"
-                    className={classnames('form-control form-control-lg', {
-                        'is-invalid': errors.password
-                    })}
-                    name="password"
-                    onChange={ this.handleInputChange }
-                    value={ this.state.password }
-                    />
-                    {errors.password && (<div className="invalid-feedback">{errors.password}</div>)}
-                </div>
-                <div className="form-group">
-                    <input
-                    type="password"
-                    placeholder="Confirm Password"
-                    className={classnames('form-control form-control-lg', {
-                        'is-invalid': errors.password_confirm
-                    })}
-                    name="password_confirm"
-                    onChange={ this.handleInputChange }
-                    value={ this.state.password_confirm }
-                    />
-                    {errors.password_confirm && (<div className="invalid-feedback">{errors.password_confirm}</div>)}
-                </div>
-                <div className="form-group">
-                {/* <Link to='/somewhere'> */}
-        //             <button  type="submit" className="btn btn-primary">
-        //                 Register User
-        //             </button> 
-        //             {/* </Link> */}
-        //         </div>
-        //     </form>
-        // </div> */}
