+++ client/src/sass/pages/_home.scss
@@ -52,3 +52,10 @@
         padding: 3rem;
     }
 }
+
+
+.section-home {
+    padding: 10rem;
+    background-color: $color-grey-light-1;
+    min-height: 100vh;
+}
